Implement GitHub-only login gate via route groups
