+++ .agents/plans/TOR-33/authentication-system-plan.md
@@ -88,9 +88,9 @@ T2 ──> T1 ──┬──> T3 ──┐
 - **location**: `apps/web/src/modules/auth/auth-client.ts`, `apps/web/src/app/api/auth/**`, `packages/api/src/modules/http/auth-plugin.ts`, `apps/server/src/**/*.test.ts`, optionally `apps/web/src/**/*.test.ts`
 - **description**: Resolve the split-host cookie problem by moving browser auth requests to the web host. Add a Next route handler under `apps/web` that proxies `/api/auth/*` to the server auth mount, preserving method, headers, body, redirect behavior, and `set-cookie` passthrough. Update `authClient` to use the web origin/same-origin auth base instead of `NEXT_PUBLIC_SERVER_URL`. Keep the Better Auth server mount unchanged on the API side. Add integration tests that cover: GitHub sign-in initiation redirect, mocked callback/session establishment through the mounted auth surface, unauthenticated session fetch, sign-out, explicit CSRF/origin rejection, and `429` on repeated sign-in initiation. Automated tests must mock the OAuth callback path; real GitHub credentials are manual-only.
 - **validation**: Browser/SSR auth traffic targets same-origin `/api/auth`; session cookies are readable by the web host; CSRF/origin checks still reject untrusted requests; DB-backed sign-in throttling returns `429`; sign-out clears the effective session.
-- **status**: In Progress
-- **log**:
-- **files edited/created**:
+- **status**: Completed
+- **log**: Replaced `/login` with a GitHub-only client entrypoint at `apps/web/src/app/(auth)/login/page.tsx`; added route-group auth guard in `apps/web/src/app/(auth)/layout.tsx` to redirect authenticated users away from `/login`; moved app root into `apps/web/src/app/(protected)/` with shared session gate in `layout.tsx` that redirects anonymous users to `/login`; removed `/dashboard` route and password-form UX files; updated `apps/web/src/modules/ui/header.tsx` to drop stale dashboard nav and `apps/web/src/modules/auth/components/user-menu.tsx` to send sign-out to `/login`. Validation run: `bun run --cwd apps/web check-types`.
+- **files edited/created**: `apps/web/src/app/(auth)/layout.tsx`, `apps/web/src/app/(auth)/login/page.tsx`, `apps/web/src/app/(protected)/layout.tsx`, `apps/web/src/app/(protected)/page.tsx`, `apps/web/src/app/dashboard/page.tsx`, `apps/web/src/app/dashboard/dashboard.tsx`, `apps/web/src/app/login/page.tsx`, `apps/web/src/app/page.tsx`, `apps/web/src/modules/auth/components/sign-in-form.tsx`, `apps/web/src/modules/auth/components/sign-up-form.tsx`, `apps/web/src/modules/auth/components/user-menu.tsx`, `apps/web/src/modules/ui/header.tsx`
 - **backlog_item_id**: `TOR-60`
 - **backlog_item_url**: `https://linear.app/stefan-projects/issue/TOR-60/tor-33-t3-make-auth-same-origin-on-the-web-host-and-verify-server-auth`
 - **relation_mode**: native
