+++ apps/web/src/app/(protected)/layout.tsx
@@ -1,0 +1,24 @@
+import { authClient } from "@/modules/auth/auth-client";
+import { headers } from "next/headers";
+import { redirect } from "next/navigation";
+
+export const dynamic = "force-dynamic";
+
+export default async function ProtectedLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  const session = await authClient.getSession({
+    fetchOptions: {
+      headers: await headers(),
+      throw: true,
+    },
+  });
+
+  if (!session?.user) {
+    redirect("/login");
+  }
+
+  return <>{children}</>;
+}
