--- apps/web/src/modules/auth/components/sign-in-form.tsx
@@ -1,150 +1,0 @@
-import { useForm } from "@tanstack/react-form";
-import { useRouter } from "next/navigation";
-import { toast } from "sonner";
-import z from "zod";
-
-import { authClient } from "@/modules/auth/auth-client";
-
-import Loader from "@/modules/ui/loader";
-import {
-  Badge,
-  Button,
-  Card,
-  CardContent,
-  CardDescription,
-  CardHeader,
-  CardTitle,
-  Input,
-  Label,
-} from "@/modules/ui/8bit";
-
-export default function SignInForm({ onSwitchToSignUp }: { onSwitchToSignUp: () => void }) {
-  const router = useRouter();
-  const { isPending } = authClient.useSession();
-
-  const form = useForm({
-    defaultValues: {
-      email: "",
-      password: "",
-    },
-    onSubmit: async ({ value }) => {
-      await authClient.signIn.email(
-        {
-          email: value.email,
-          password: value.password,
-        },
-        {
-          onSuccess: () => {
-            router.push("/dashboard");
-            toast.success("Sign in successful");
-          },
-          onError: (error) => {
-            toast.error(error.error.message || error.error.statusText);
-          },
-        },
-      );
-    },
-    validators: {
-      onSubmit: z.object({
-        email: z.email("Invalid email address"),
-        password: z.string().min(8, "Password must be at least 8 characters"),
-      }),
-    },
-  });
-
-  if (isPending) {
-    return <Loader />;
-  }
-
-  return (
-    <Card className="mx-auto mt-10 w-full max-w-xl">
-      <CardHeader>
-        <Badge className="bg-primary/14 text-primary-foreground">VERIFICATION</Badge>
-        <CardTitle className="text-base md:text-xl">IDENTITY REQUIRED</CardTitle>
-        <CardDescription>
-          Submit credentials to regain terminal access. The system does not recognize unauthorized presences.
-        </CardDescription>
-      </CardHeader>
-      <CardContent>
-        <form
-          onSubmit={(e) => {
-            e.preventDefault();
-            e.stopPropagation();
-            form.handleSubmit();
-          }}
-          className="space-y-5"
-        >
-          <div>
-            <form.Field name="email">
-              {(field) => (
-                <div className="space-y-2">
-                  <Label htmlFor={field.name}>EMAIL</Label>
-                  <Input
-                    id={field.name}
-                    name={field.name}
-                    type="email"
-                    value={field.state.value}
-                    onBlur={field.handleBlur}
-                    onChange={(e) => field.handleChange(e.target.value)}
-                  />
-                  {field.state.meta.errors.map((error) => (
-                    <p key={error?.message} className="retro-copy text-sm text-destructive">
-                      {error?.message}
-                    </p>
-                  ))}
-                </div>
-              )}
-            </form.Field>
-          </div>
-
-          <div>
-            <form.Field name="password">
-              {(field) => (
-                <div className="space-y-2">
-                  <Label htmlFor={field.name}>PASSWORD</Label>
-                  <Input
-                    id={field.name}
-                    name={field.name}
-                    type="password"
-                    value={field.state.value}
-                    onBlur={field.handleBlur}
-                    onChange={(e) => field.handleChange(e.target.value)}
-                  />
-                  {field.state.meta.errors.map((error) => (
-                    <p key={error?.message} className="retro-copy text-sm text-destructive">
-                      {error?.message}
-                    </p>
-                  ))}
-                </div>
-              )}
-            </form.Field>
-          </div>
-
-          <form.Subscribe
-            selector={(state) => ({
-              canSubmit: state.canSubmit,
-              isSubmitting: state.isSubmitting,
-            })}
-          >
-            {(state) => (
-              <Button
-                type="submit"
-                className="w-full"
-                disabled={!state.canSubmit || state.isSubmitting}
-              >
-                {state.isSubmitting ? "VERIFYING..." : "AUTHENTICATE"}
-              </Button>
-            )}
-          </form.Subscribe>
-        </form>
-
-        <div className="mt-5 flex flex-wrap items-center justify-between gap-3">
-          <p className="retro-copy text-sm text-muted-foreground">No credentials registered?</p>
-          <Button variant="outline" size="sm" onClick={onSwitchToSignUp}>
-            Create Record
-          </Button>
-        </div>
-      </CardContent>
-    </Card>
-  );
-}
