+++ apps/web/src/modules/auth/components/user-menu.tsx
@@ -47,7 +47,7 @@ export default function UserMenu() {
               authClient.signOut({
                 fetchOptions: {
                   onSuccess: () => {
-                    router.push("/");
+                    router.push("/login");
                   },
                 },
               });
